Reading invoice data from CreateInvoinceSRWindow
This window collects all necessary data to create an invoice and stores it in the global variable 'invoice'.
The collected information is used to send package details to the server.

Data includes:
- pickup address
- delivery address
- recipient info
- sender info
- package comment

The AddRecieverInCreatingInvoinceWindow returns a 'RecipientData' object containing all recipient contact details for the courier.

The Invoice class is currently implemented only for this window.